requirements.txt
streamlit
pandas
sweetviz
chardet
seaborn
matplotlib
openpyxl
google-generativeai
